+++ picle/__init__.py
@@ -1,4 +1,3 @@
 from .picle import App
-from .cache import Cache
 
-__all__ = ["App", "Cache"]
+__all__ = ["App"]